+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 * Renamed optional-requirements to nvidia-requirements to clarify its purpose
 * Installer shows a progress indicator rather than displaying unwanted clutter from apt / pip
 * Passwords are no longer echoed back in the installer
+* Installer prompts are prefilled with default values where appropriate
 
 1.0.0